+++ adjustments.f08
@@ -48,13 +48,13 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_channel_velocity_limiter'
 
-        real,      target, intent(in out)  :: elemR(:,:)
+        real(8),      target, intent(in out)  :: elemR(:,:)
         logical,           intent(in out)  :: elemYN(:,:)
         integer,           intent(in)      :: elemI(:,:)
         integer,           intent(in)      :: ei_elem_type, elemType, eYN_IsAdhocFlowrate
         integer,           intent(in)      :: er_Velocity_new
 
-        real,  pointer ::  velocity(:)
+        real(8),  pointer ::  velocity(:)
 
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
@@ -90,13 +90,13 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_pipe_velocity_limiter'
 
-        real,      target, intent(in out)  :: elemR(:,:)
+        real(8),      target, intent(in out)  :: elemR(:,:)
         logical,           intent(in out)  :: elemYN(:,:)
         integer,           intent(in)      :: elemI(:,:)
         integer,           intent(in)      :: ei_elem_type, elemType, eYN_IsAdhocFlowrate
         integer,           intent(in)      :: er_Velocity_new
 
-        real,  pointer ::  velocity(:)
+        real(8),  pointer ::  velocity(:)
 
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
@@ -135,10 +135,10 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_junction_branch_velocity_limit'
 
-        real,  target,     intent(in out)  :: elemMR(:,:)
+        real(8),  target,     intent(in out)  :: elemMR(:,:)
         integer,           intent(in)      :: elemMI(:,:)
 
-        real,  pointer :: velocity(:), flowrate(:), area(:)
+        real(8),  pointer :: velocity(:), flowrate(:), area(:)
 
         integer :: mm
 
@@ -187,14 +187,14 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_face_dynamic_limits'
 
-        real,                  intent(in out)  :: faceR(:,:)
+        real(8),                  intent(in out)  :: faceR(:,:)
         ! note, requires that elemR be provided separately for upstream and downstream
         ! these can be the samy arrays (e.g. in the case of elem2R for channel-channel)
-        real,                  intent(in)      :: volumeUp(:), volumeDn(:)
+        real(8),                  intent(in)      :: volumeUp(:), volumeDn(:)
         integer,   target,     intent(in)      :: faceI(:,:)
         logical,               intent(in)      :: facemask(:), Ltemp
 
-        real,  pointer :: volFrac
+        real(8),  pointer :: volFrac
         integer,   pointer :: eUp(:), eDn(:)
 
         !--------------------------------------------------------------------------
@@ -271,12 +271,12 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_for_zero_geometry'
 
-        real,          intent(in out)  ::  elem2R(:,:)
-        real, target,  intent(in out)  ::  elemMR(:,:)
+        real(8),          intent(in out)  ::  elem2R(:,:)
+        real(8), target,  intent(in out)  ::  elemMR(:,:)
         integer,       intent(in)      ::  elemMI(:,:)
         logical,       intent(in)      ::  elem2YN(:,:), elemMYN(:,:)
 
-        real, pointer  :: area(:), topwidth(:)
+        real(8), pointer  :: area(:), topwidth(:)
 
         integer :: mm
 
@@ -318,7 +318,7 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_negative_volume_reset'
 
-        real,  intent(in out)  ::  volume(:)
+        real(8),  intent(in out)  ::  volume(:)
 
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
@@ -349,7 +349,7 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_negative_area_reset'
 
-        real,  intent(in out)  ::  area(:)
+        real(8),  intent(in out)  ::  area(:)
 
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
@@ -380,8 +380,8 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_negative_eta_reset'
 
-        real,       intent(in out)  ::  eta(:)
-        real,       intent(in out)  ::  zbottom(:)
+        real(8),       intent(in out)  ::  eta(:)
+        real(8),       intent(in out)  ::  zbottom(:)
         logical,    intent(in)      ::  maskarray(:)
 
         !--------------------------------------------------------------------------
@@ -416,7 +416,7 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_smallvolumes'
 
-        real,      intent(in out)      :: elem2R(:,:),  elemMR(:,:)
+        real(8),      intent(in out)      :: elem2R(:,:),  elemMR(:,:)
         integer,   intent(in)          :: elem2I(:,:),  elemMI(:,:)
         logical,   intent(in out)      :: elem2YN(:,:), elemMYN(:,:)
         integer,   intent(in)          :: e2r_VolumeColumn, eMr_VolumeColumn
@@ -449,19 +449,19 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_Vshaped_flowrate'
 
-        real,      target,     intent(in out)  :: elem2R(:,:)
-        real,      target,     intent(in)      :: faceR(:,:)
+        real(8),      target,     intent(in out)  :: elem2R(:,:)
+        real(8),      target,     intent(in)      :: faceR(:,:)
         integer,   target,     intent(in)      :: elem2I(:,:)
         logical,   target,     intent(in out)  :: elem2YN(:,:)
 
-        real,      pointer     :: elemFlow(:), faceFlow(:), elemAdjust(:), elemVel(:), elemArea(:)
-        real,      pointer     :: tscaleUp(:), tscaleDn(:)
+        real(8),      pointer     :: elemFlow(:), faceFlow(:), elemAdjust(:), elemVel(:), elemArea(:)
+        real(8),      pointer     :: tscaleUp(:), tscaleDn(:)
         integer,   pointer     :: mapUp(:), mapDn(:)
         logical,   pointer     :: elemMask(:)
 
         integer :: e2r_adjustflow, e2YN_mask
 
-        real,      pointer     :: coef
+        real(8),      pointer     :: coef
 
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
@@ -549,19 +549,19 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_pressure_in_pipe'
 
-        real,      target,     intent(in out)  :: elem2R(:,:)
-        real,      target,     intent(in)      :: faceR(:,:)
+        real(8),      target,     intent(in out)  :: elem2R(:,:)
+        real(8),      target,     intent(in)      :: faceR(:,:)
         integer,   target,     intent(in)      :: elem2I(:,:)
         logical,   target,     intent(in out)  :: elem2YN(:,:)
 
-        real,      pointer     :: elemEta(:), elemAdjust(:), faceEtaUp(:)
-        real,      pointer     :: tscaleUp(:), tscaleDn(:), zCrown(:), faceEtaDn(:)
+        real(8),      pointer     :: elemEta(:), elemAdjust(:), faceEtaUp(:)
+        real(8),      pointer     :: tscaleUp(:), tscaleDn(:), zCrown(:), faceEtaDn(:)
         integer,   pointer     :: mapUp(:), mapDn(:), elemType(:)
         logical,   pointer     :: elemMask(:)
 
         integer :: e2r_adjustEta, e2YN_mask
 
-        real,      pointer     :: coef
+        real(8),      pointer     :: coef
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
 
@@ -650,7 +650,7 @@ contains
         !
         character(64) :: subroutine_name = 'adjust_zero_velocity_at_zero_volume'
 
-        real,      intent(in out)  :: elem2R(:,:),  elemMR(:,:)
+        real(8),      intent(in out)  :: elem2R(:,:),  elemMR(:,:)
         logical,   intent(in out)  :: elem2YN(:,:), elemMYN(:,:)
         integer,   intent(in)      :: e2r_VelocityColumn, e2r_VolumeColumn
         integer,   intent(in)      :: eMr_VelocityColumn, eMr_VolumeColumn
@@ -684,12 +684,12 @@ contains
         !
         character(64) :: subroutine_name = 'smallvolume_identification'
 
-        real,    target,   intent(in out)  :: elem2R(:,:),  elemMR(:,:)
+        real(8),    target,   intent(in out)  :: elem2R(:,:),  elemMR(:,:)
         integer, target,   intent(in)      :: elem2I(:,:),  elemMI(:,:)
         logical, target,   intent(in out)  :: elem2YN(:,:), elemMYN(:,:)
         integer,           intent(in)      :: eTr_Volume2,  eTr_VolumeM
 
-        real,      pointer :: smallvolumeratio(:), smallvolume(:), tvolume(:)
+        real(8),      pointer :: smallvolumeratio(:), smallvolume(:), tvolume(:)
         integer,   pointer :: elemtype(:)
 
         logical, pointer :: issmallvolume(:)
@@ -760,8 +760,8 @@ contains
         !
         character(64) :: subroutine_name = 'smallvolume_identification_for_element'
 
-        real,      intent(in out)  :: volume(:), smallvolumeratio(:)
-        real,      intent(in)      :: smallvolume(:)
+        real(8),      intent(in out)  :: volume(:), smallvolumeratio(:)
+        real(8),      intent(in)      :: smallvolume(:)
         integer,   intent(in)      :: elemtype(:)
         logical,   intent(in out)  :: issmallvolume(:)
         integer,   intent(in)      :: thiselementtype
@@ -807,13 +807,13 @@ contains
         !
         character(64) :: subroutine_name = 'smallvolume_geometry'
 
-        real,              intent(in out)  :: elem2R(:,:)
+        real(8),              intent(in out)  :: elem2R(:,:)
         integer,           intent(in)      :: elem2I(:,:)
-        real,      target, intent(in out)  :: elemMR(:,:)
+        real(8),      target, intent(in out)  :: elemMR(:,:)
         integer,           intent(in)      :: elemMI(:,:)
         logical,           intent(in)      :: elem2YN(:,:), elemMYN(:,:)
 
-        real,  pointer :: area(:), topwidth(:)
+        real(8),  pointer :: area(:), topwidth(:)
         integer :: mm
 
         !--------------------------------------------------------------------------
@@ -873,7 +873,7 @@ contains
         !
         character(64) :: subroutine_name = 'smallvolume_element_geometry_reset'
 
-        real,      target, intent(in out)  :: elemR(:,:)
+        real(8),      target, intent(in out)  :: elemR(:,:)
         integer,   target, intent(in)      :: elemI(:,:)
 
         logical,   intent(in)  :: elemYN(:,:)
@@ -883,9 +883,9 @@ contains
         integer,   intent(in)  :: er_HydDepth, er_HydRadius, er_Topwidth
         integer,   intent(in)  :: eYN_IsSmallVolume
 
-        real, pointer  ::  area(:), eta(:), perimeter(:), zbottom(:)
-        real, pointer  ::  hyddepth(:), hydradius(:)
-        real, pointer  ::  topwidth(:)
+        real(8), pointer  ::  area(:), eta(:), perimeter(:), zbottom(:)
+        real(8), pointer  ::  hyddepth(:), hydradius(:)
+        real(8), pointer  ::  topwidth(:)
 
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
@@ -923,14 +923,14 @@ contains
         !
         character(64) :: subroutine_name = 'smallvolume_junctionbranch_reset'
 
-        real,      target, intent(in out)  :: elemMR(:,:)
+        real(8),      target, intent(in out)  :: elemMR(:,:)
         logical,           intent(in)      :: elemMYN(:,:)
 
         ! dnstream_face_per_elemM or upstream_face_per_elemM
         integer,           intent(in)      :: Dir_face_per_elem
         integer,           intent(in)      :: eMr_AreaDir(:), eMr_TopwidthDir(:)
 
-        real,  pointer :: area(:), topwidth(:)
+        real(8),  pointer :: area(:), topwidth(:)
         integer        :: mm
 
         !--------------------------------------------------------------------------
@@ -968,15 +968,15 @@ contains
         !
         character(64) :: subroutine_name = 'reset_element_for_zero_values'
 
-        real,  target,  intent(in out) :: elemR(:,:)
+        real(8),  target,  intent(in out) :: elemR(:,:)
         logical,        intent(in)     :: elemYN(:,:)
 
         integer,        intent(in)     :: er_Area, er_Eta, er_Zbottom, er_Topwidth
         integer,        intent(in)     :: er_Perimeter, er_HydDepth, er_HydRadius
         integer,        intent(in)     :: eYN_IsSmallVolume
 
-        real,  pointer :: area(:), eta(:), topwidth(:), perimeter(:)
-        real,  pointer :: zbottom(:), hyddepth(:), hydradius(:)
+        real(8),  pointer :: area(:), eta(:), topwidth(:), perimeter(:)
+        real(8),  pointer :: zbottom(:), hyddepth(:), hydradius(:)
 
         integer :: ii
 
@@ -1060,12 +1060,12 @@ contains
         !
         character(64) :: subroutine_name = 'reset_juctionbranches_for_zero_values'
 
-        real,      target,     intent(in out)  :: elemMR(:,:)
+        real(8),      target,     intent(in out)  :: elemMR(:,:)
         integer,               intent(in)      :: elemMI(:,:)
         logical,               intent(in)      :: elemMYN(:,:)
         integer,               intent(in)      :: Dir_face_per_elemM, eMr_nfaces_Dir
         integer,               intent(in)      :: eMr_AreaDir(:), eMr_TopwidthDir(:)
-        real,  pointer :: area(:), topwidth(:)
+        real(8),  pointer :: area(:), topwidth(:)
         integer        :: mm
 
         !--------------------------------------------------------------------------
@@ -1110,11 +1110,11 @@ contains
         !
         character(64) :: subroutine_name = 'zero_velocity_at_zero_volume'
 
-        real,      target,     intent(in out)  :: elemR(:,:)
+        real(8),      target,     intent(in out)  :: elemR(:,:)
         logical,               intent(in out)  :: elemYN(:,:)
         integer,               intent(in)      :: er_Velocity, er_Volume, eYN_IsAdhocFlowrate
 
-        real,  pointer :: velocity(:),  volume(:)
+        real(8),  pointer :: velocity(:),  volume(:)
 
         !--------------------------------------------------------------------------
         if ((debuglevel > 0) .or. (debuglevelall > 0)) print *, '*** enter ',subroutine_name
